// frontend/src/features/team/types/contracts/user.ts
import type { UserRole } from "../user";

// I haven't put id here because it will be generated by the backend when a new user is created.
export interface CreateUserRequest {
  name: string;
  avatar?: string;
  role: UserRole;
}

export interface UpdateUserRequest {
  name: string;
  avatar?: string;
  role: UserRole;
}